requirements.txt
apache-airflow>=2.1.2
requests>=2.26.0
boto3>=1.18.21
python-dotenv>=0.19.0